Move to_utc_timestamp to the audit module
It's only used there, and let's not make it as blessed as it currently
is, let's just have it more private for now
